Renamed and moved qrcode wallet dialog to template
