+++ layouts/auth.vue
@@ -5,3 +5,4 @@
     </v-content>
   </v-app>
 </template>
+
