+++ store/wallet.js
@@ -3,7 +3,8 @@ import { sleep } from '@/lib/utils'
 export const state = () => ({
   loading: false,
   wallets: [],
-  selectedWallet: null
+  selectedWallet: null,
+  popupAddress: false
 })
 
 export const getters = {
@@ -25,6 +26,9 @@ export const getters = {
   isLoggedIn: state => {
     return state.selectedWallet === null ? false : true
   },
+  showPopupAddress: state => {
+    return state.popupAddress
+  }
 }
 
 export const mutations = {
@@ -44,6 +48,9 @@ export const mutations = {
   },
   disconnect: (state, index) => {
     state.selectedWallet = null
+  },
+  tooglePopup: (state) => {
+    state.popupAddress = !state.popupAddress
   }
 }
